feat: :sparkles: Ajustes página de login
Realizado ajustes na página de login e verificações

#39
