+++ src/css/global/global.css
@@ -9,7 +9,7 @@
 body {
   margin: 0;
   font-family: 'Roboto Flex', sans-serif;
-  color: var(--cor-b);
+  color: var(--cor-w);
 }
 
 img {
