+++ src/js/scriptLogin.js
@@ -1,5 +1,4 @@
-const head = document.querySelector('head');
-
+import { head } from "./script.js";
 const loginController = document.createElement("script");
 const loginView = document.createElement("script");
